move logic to Item
